requirements.txt
redis~=5.2.0
aiogram~=3.14.0
pydantic~=2.9.2
requests~=2.32.3
python-dotenv~=1.0.1